make confxml template handle boolean variables cleaner
Some settings in conf.xml must be set to literal 'yes' or 'no' instead
of 'true' or 'false'. But from ansible they could be interpreted as
boolean. This leads to errors as 'wrong' setting strings seem to get
silently ignored by existdb.
Now we cast those variables to boolean, test for true and emit 'yes'
or 'no'.